# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Coopvest Africa Admin Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: getMembers
      tags: [members]
      summary: List all members
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended, pending]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembersListResponse"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberBody"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    put:
      operationId: updateMember
      tags: [members]
      summary: Update member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberBody"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsListResponse"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Send notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationBody"
      responses:
        "201":
          description: Created notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /support/tickets:
    get:
      operationId: getSupportTickets
      tags: [support]
      summary: List support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketsListResponse"
  /support/tickets/{id}:
    get:
      operationId: getSupportTicket
      tags: [support]
      summary: Get support ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Support ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
components:
  schemas:
...
    Member:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended, pending]
        joinDate:
          type: string
          format: date
        totalContributions:
          type: number
        activeLoan:
          type: number
        riskScore:
          type: integer
        address:
          type: string
        occupation:
          type: string
        avatarInitials:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, memberId, firstName, lastName, email, phone, status, joinDate, totalContributions, activeLoan, riskScore, createdAt]
    MembersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateMemberBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        occupation:
          type: string
      required: [firstName, lastName, email, phone]
    UpdateMemberBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended, pending]
        address:
          type: string
        occupation:
          type: string
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, success, error, alert]
        isRead:
          type: boolean
        targetAudience:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, message, type, isRead, createdAt]
    NotificationsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, unreadCount, page, limit]
    CreateNotificationBody:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, success, error, alert]
        targetAudience:
          type: string
      required: [title, message, type]
...
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: string
        memberId:
          type: integer
        memberName:
          type: string
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        category:
          type: string
        assignedTo:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
      required: [id, ticketId, memberId, memberName, subject, status, priority, createdAt]
    SupportTicketsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SupportTicket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
...