# lib/api-zod/src/generated/types/membersListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Coopvest Africa Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Member } from "./member";

export interface MembersListResponse {
  data: Member[];
  total: number;
  page: number;
  limit: number;
}